+++ CHANGES.md
@@ -3,3 +3,7 @@
 ## Changes for v0.0.3-beta.1:
 - Added command builder tests.
 - Improved command builder class.
+
+## Changes for v0.0.3-beta.2:
+- Builder tests completed.
+- Added services tests.